refactor: move modals file to a more accessible path
